+++ inst/app/tabs/load.R
@@ -1,8 +1,12 @@
 ### load_tab ----
 load_tab <- tabItem(
   tabName = "load_tab",
-  p("This app is under development; all materials created should be carefully checked."),
-  fileInput("load_xml", "Load from XML", multiple = TRUE, width = "100%", accept = ".xml"),
+  p("This app is under development; all materials created should be carefully checked.", class = "warning"),
+  p("Use `pdf2grobid()` to create XML files from PDFs. This relies on an external or local grobid server, and will save files to your machine, so needs to run outside this app."),
+  fileInput("load_xml", "Load XML",
+            multiple = TRUE,
+            width = "100%",
+            accept = ".xml"),
   textOutput("n_papers_loaded"),
   box(width = 12, collapsible = TRUE, collapsed = FALSE,
       title = "Paper Info",
@@ -10,9 +14,6 @@ load_tab <- tabItem(
       uiOutput("paper_title"),
       uiOutput("paper_desc"),
       textOutput("paper_keywords")
-      # textInput("paper_title", "paper Title", "", "100%"),
-      # textAreaInput("paper_desc", "paper Description", "", "100%"),
-      # textInput("paper_keywords", "Keywords (separate with semicolons)", "", "100%", )
   )
 )